requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.6
future==0.18.2
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
